Added a Client class
I added a client class where the host and their respctive remote port can be parsed through the constructor parameters. Added a method called StartClient(); which will be called when the user wants to connect to a specific host through a specific remote port. Added a Server class, but it does not have anything in it right now