Added a multithreaded solution to the Server, allowing multiple clients to connect simultaneously
Added a method to the server class to check whether the remote port given is available. Created a new class called ServerConnection which implements the Runnable interface. This class is responsible for acceepting new client connections and storing the connected client's socket in a ArrayList. The ServerConnection class runs in a seperate thread to not block other code in the Server class