Added the ability for a client to send messages and support for the server to receive client messages
Added a new ServerThread class in the Server project. Each connected client will have a seperate thread, these threads will check wether the client has sent a message and will display these messages in the console. In the future it will also have support to relay these messages.

Added a new SendMessageThread for the Client project.  This class is responsible for client users to type messages and send the typed message it to the server.

Made qucik changes to both the Server and Client classes to adapt to the new classes.