Added a new class to debug the server while running, added support to remove not currently used sockets
Added a HashMap, where the Keys are Sockets and Values are Threads. This is used to keep track of which socket is associated with a Thread.

Added a new DebugServerThread class , used to be able to debug the server while running. This class implments the  Runnable interface and uses direct input from the console. The only function which is available is displaying each Socket's port present in the socketList.

Added support to remove sockets which is not being used to host a client, because of disconnect or other failur. The code can remove sockets from the socketList but cannot cancel the thread running the sockets, which will be addressed in the coming future.