Moved the class DebugServerThread
In previous versions of the program the class called DebugServerThread was positioned inside the Server.java file. SocketList was changed from private to package-private, then this allowed the DebugServerThread to be called from a sperate file. Now a new file has been added called DebugServerThread.java which consist of the class DebugServerThread.