Added the ability to send messages to other connected clients
Added a new class to the Chat program called ReceiveMessagesThread. This class is used to read incoming messages on the socket's input stream. As the name indicates, it is another thread and is started in the Client class.

Added a new class to the Server application called SendMessagesThread. This class is responsible for sending out client messages which has been queued in a new ConcurrentLinkedQueue. This class is not quite working as intended and is currently provoking something that makes ServerThreads shutdown. The way that it is supposed to work is the following: check if any messages is currently available in the messageQueue, if that is true, send it to each connected socket’s output stream.

In the Server class, a new method has been added called "ReceiveMessages". This method can be called by a ServerThreads, which can then enter their respective client's messages. This method will then take the message and add it to the message queue.
In the UpdateServer method, another block of code has been added to substitute the SendMessagesThread class. This block of code checks the messageQueue if the queue is currently empty, if it is not, it will take that message and print in to each connected client socket's output stream. When the message has been sent, it will be removed from the queue.