Clients can send UDP requests with their username for approval
A client must send a UDP request with their username to the server, if a user has alreay taken that username the client will be denied to connect.  This adds a new verification step to get into the server. This also stops the socket used for reconnection attempts to appear as a user in the server